Move TestIsAbsoluteURL and TestGetAbsoluteURL to test_utils.py
Toward #2628
