+++ refinery/core/tests.py
@@ -11,7 +11,7 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.core.management import CommandError, call_command
 from django.db import connection
 from django.db.migrations.executor import MigrationExecutor
-from django.test import TestCase, override_settings
+from django.test import TestCase
 from django.utils import timezone
 
 from guardian.core import ObjectPermissionChecker
@@ -39,9 +39,8 @@ from .models import (INPUT_CONNECTION, OUTPUT_CONNECTION, Analysis,
                      UserProfile, Workflow, WorkflowDataInputMap,
                      WorkflowEngine, invalidate_cached_object)
 from .search_indexes import DataSetIndex
-from .utils import (filter_nodes_uuids_in_solr, get_absolute_url,
-                    get_aware_local_time, get_resources_for_user,
-                    is_absolute_url, move_obj_to_front,
+from .utils import (filter_nodes_uuids_in_solr, get_aware_local_time,
+                    get_resources_for_user, move_obj_to_front,
                     which_default_read_perm)
 from .views import AnalysesViewSet, DataSetsViewSet
 
@@ -2296,49 +2295,3 @@ class TestManagementCommands(TestCase):
                 str(galaxy_instance.id),
                 "non-existent group name"
             )
-
-
-class TestIsAbsoluteURL(TestCase):
-
-    def test_is_absolute_url(self):
-        self.assertTrue(is_absolute_url('http://example.org'))
-        self.assertTrue(is_absolute_url('ftp://example.org'))
-
-    def test_is_not_absolute_url(self):
-        self.assertFalse(is_absolute_url('/path/to/file'))
-        self.assertFalse(is_absolute_url('example.org'))
-
-    def test_is_absolute_url_with_empty_input(self):
-        self.assertFalse(is_absolute_url(None))
-        self.assertFalse(is_absolute_url(''))
-
-    def test_get_absolute_url_with_absolute_url(self):
-        self.assertEqual(get_absolute_url('http://example.org'),
-                         'http://example.org')
-
-
-@override_settings(REFINERY_URL_SCHEME='http')
-class TestGetAbsoluteURL(TestCase):
-
-    def setUp(self):
-        self.current_site = Site.objects.get_or_create(name='Test',
-                                                       domain='example.org')[0]
-
-    def test_get_absolute_url_with_relative_url(self):
-        with override_settings(SITE_ID=self.current_site.id):
-            self.assertEqual(get_absolute_url('/path/to/file'),
-                             'http://example.org/path/to/file')
-
-    def test_get_absolute_url_with_absolute_url(self):
-        with override_settings(SITE_ID=self.current_site.id):
-            self.assertEqual(get_absolute_url('http://example.org'),
-                             'http://example.org')
-
-    def test_get_absolute_url_with_empty_input(self):
-        with override_settings(SITE_ID=self.current_site.id):
-            self.assertEqual(get_absolute_url(None), None)
-            self.assertEqual(get_absolute_url(''), '')
-
-    def test_get_absolute_url_with_invalid_current_site(self):
-        Site.objects.all().delete()
-        self.assertIsNone(get_absolute_url('/path/to/file'))
